Get MS working with gatsby
