+++ gatsby/src/templates/MishnaStudy/Mishnayot/Mishnayot.jsx
