+++ gatsby/src/templates/MishnaStudy/Mishnayot/TeacherCard.jsx
